Update PRs in CoinFabrik_On_Ink_Integration_Tests-2.md
We explain below two changes that we had to perform to ensure that our PRs passed the corresponding checks, and to be consistent with new developments we have performed on other unimplemented functions in integration tests.

We initially performed the Pull Request [#1963](https://github.com/paritytech/ink/pull/1963) to implement instantiate_contract(), but decided to close it and perform a new Pull Request [#1988](https://github.com/paritytech/ink/pull/1988) to include necessary implementations of other related functions, which were developed as a subsequent milestone to this Milestone 2. These functions were all related to each other, and a single PR seemed like the tidiest option.

As part of an ad-hoc development in PR [#58] (https://github.com/paritytech/cargo-contract/pull/1358), adding support to workspaces, we were requested to ensure that ink-examples worked correctly with this implementation. We initially performed the Pull Request [#44](https://github.com/paritytech/ink-examples/pull/44) to show this. However, we had to close this PR because some modifications needed to be performed for checks to pass. We made these changes in Pull Request [#52](https://github.com/paritytech/ink-examples/pull/52).

We updated the Milestone Report and Table of Deliveries to reflect these final changes.